+++ ZapLibUnitTest/ImplicitMailerTests.cs
@@ -20,11 +20,11 @@ namespace ZapLib.Tests
             string MAIL_ACT = "support5@iqs-t.com";
             string MAIL_PWD = "";
             */
-            
+
             string SMTP_HOST = "smtp.gmail.com";
             string MAIL_ACT = "maryzap123@gmail.com";
             string MAIL_PWD = "";
-            
+
 
             int MAIL_PORT = 465;
             bool MAIL_SSL = true;
@@ -34,8 +34,10 @@ namespace ZapLib.Tests
             string TO = "zaplin@iqs-t.com,sollin@iqs-t.com";
             string SUBJECT = $"This is a test mail (ZapLib.ImplicitMailer) - {DateTime.Now}";
             string BODY = $"<h1>Test Mail by ZapLib ImplicitMailer</h1>";
+            string[] Attachment = new string[] { @"D:\Downloads\delivered.png", @"D:\Downloads\abstract-christmas-tree.zip", @"D:\Downloads\.gitignore" };
 
-            bool result = m.Send(TO, SUBJECT, BODY);
+
+            bool result = m.Send(TO, SUBJECT, BODY, attchments: Attachment);
             Trace.WriteLine(m.ErrMsg);
             Assert.IsTrue(result);
         }